Port to Acorn 6.0 plugin interface
