+++ README.md
@@ -17,44 +17,24 @@ Please note that this tool only parses source code to JSX AST, which is useful f
 
 ## Usage
 
-You can use module directly in order to get Acorn instance with plugin installed:
+Requiring this module provides you with an Acorn plugin that you can use like this:
 
 ```javascript
-var acorn = require('acorn-jsx');
-```
-
-Or you can use `inject.js` for injecting plugin into your own version of Acorn like following:
-
-```javascript
-var acorn = require('acorn-jsx/inject')(require('./custom-acorn'));
-```
-
-Then, use `plugins` option whenever you need to support JSX while parsing:
-
-```javascript
-var ast = acorn.parse(code, {
-  plugins: { jsx: true }
-});
+var acorn = require("acorn");
+var jsx = require("acorn-jsx");
+acorn.Parser.extend(jsx()).parse("my(<jsx/>, 'code');");
 ```
 
 Note that official spec doesn't support mix of XML namespaces and object-style access in tag names (#27) like in `<namespace:Object.Property />`, so it was deprecated in `acorn-jsx@3.0`. If you still want to opt-in to support of such constructions, you can pass the following option:
 
 ```javascript
-var ast = acorn.parse(code, {
-  plugins: {
-    jsx: { allowNamespacedObjects: true }
-  }
-});
+acorn.Parser.extend(jsx({ allowNamespacedObjects: true }))
 ```
 
 Also, since most apps use pure React transformer, a new option was introduced that allows to prohibit namespaces completely:
 
 ```javascript
-var ast = acorn.parse(code, {
-  plugins: {
-    jsx: { allowNamespaces: false }
-  }
-});
+acorn.Parser.extend(jsx({ allowNamespaces: false }))
 ```
 
 Note that by default `allowNamespaces` is enabled for spec compliancy.
