+++ package.json
@@ -18,11 +18,10 @@
   "scripts": {
     "test": "node test/run.js"
   },
-  "dependencies": {
-    "acorn": "^5.0.3"
+  "peerDependencies": {
+    "acorn": "^6.0.0"
   },
   "devDependencies": {
-    "chai": "^3.0.0",
-    "mocha": "^3.3.0"
+    "acorn": "^6.0.0"
   }
 }
