+++ test/run.js
@@ -17,10 +17,16 @@ function log(title, message) {
   if (typeof console === "object") console.log(title, message);
 }
 
+const acorn = require("acorn"), jsx = require("..")
+const Parser = acorn.Parser.extend(jsx())
+
 var stats, modes = {
   Normal: {
     config: {
-      parse: require("..").parse
+      parse: (input, options, pluginOptions) => {
+        if (!pluginOptions) return Parser.parse(input, options)
+        else return acorn.Parser.extend(jsx(pluginOptions)).parse(input, options)
+      }
     }
   }
 };
